+++ src/bin/loom.rs
@@ -1,66 +1,70 @@
 extern crate getopts;
 extern crate loom;
-
-use loom::net;
-use loom::state;
-use loom::gossip;
-use loom::data;
+extern crate rpassword;
 
 use getopts::Options;
 use std::env;
 use std::string::String;
+
+use loom::wallet::{EncryptedWallet, Wallet};
 
 fn print_usage(program: &str, opts: Options) {
     let brief = format!("Usage: {} FILE [options]", program);
     print!("{}", opts.usage(&brief));
 }
 
-fn loom(port: u16) {
-    let mut num = 0;
-    let start = num;
-    net::read(&srv, &mut m[start..], &mut num).expect("read");
-    let end = num;
-    s.execute(&mut m[start..end]).expect("state");
-    g.execute(&mut m[start..end]).expect("gossip");
-    for s in &g.subs {
-        net::sendtov4(&srv, &m[start..end], &mut num, s.addr, s.port).expect("send");
-    }
+fn new_key_pair() {
+    let path = "loom.wallet";
+    let prompt = "./loom.wallet password: ";
+    let pass = rpassword::prompt_password_stdout(prompt).unwrap();
+    let ew = EncryptedWallet::from_file(path).unwrap_or(EncryptedWallet::new());
+    let mut w = Wallet::decrypt(ew, pass.as_bytes()).expect("decrypt");
+    let kp = Wallet::new_keypair();
+    w.add_key_pair(kp);
+    w.encrypt(pass.as_bytes())
+        .expect("encrypt")
+        .to_file(path)
+        .expect("write");
 }
 
-fn spool(loom: str) {
-    let mut num = 0;
-    let start = num;
-    net::read(&srv, &mut m[start..], &mut num).expect("read");
-    let end = num;
-    s.execute(&mut m[start..end]).expect("state");
-    g.execute(&mut m[start..end]).expect("gossip");
-}
+fn transfer(_from: String, _to: String, _amnt: u64) {}
+
+fn balance(_addr: String) {}
 
 pub fn main() {
-    let srv = net::server().expect("server");
-    let mut s = state::State::new(1024);
-    let mut g = gossip::Gossip::new(1024);
-    let mut m = vec![data::Message::default(); 1024];
+    let args: Vec<String> = env::args().collect();
+    let program = args[0].clone();
 
+    let mut opts = Options::new();
+    opts.optflag("c", "", "create a new address");
+    opts.optflag("x", "", "transfer");
+    opts.optflag("b", "", "check the balance of destination address");
+    opts.optflag("l", "list", "list your addresses and balances");
+    opts.optflag("h", "help", "print this help menu");
+    opts.optopt("t", "", "destination address", "ADDRESS");
+    opts.optopt("f", "", "source address", "ADDRESS");
+    opts.optopt("a", "", "amount", "AMOUNT");
+    let matches = match opts.parse(&args[1..]) {
+        Ok(m) => m,
+        Err(f) => panic!(f.to_string()),
+    };
     if matches.opt_present("h") {
         print_usage(&program, opts);
         return;
     }
-
-    opts.optopt(
-        "s",
-        "",
-        "Run as a Spool node with the Loom address",
-        "ADDRESS",
-    );
-    opts.optopt("l", "", "Run as a Loom with a listen port", "PORT");
-
-    if matches.opt_str("s").is_some() {
-        let loom = matches.opt_str("s").expect("missing loom address");
-        spool(loom);
-    } else {
-        let ports = matches.opt_str("l").expect("missing loom port");
-        let port = ports.parse("expecting u16 number for port");
-        loom(port);
+    if matches.opt_present("c") {
+        new_key_pair();
+        return;
+    } else if matches.opt_present("x") {
+        let to = matches.opt_str("t").expect("missing destination address");
+        let from = matches.opt_str("f").expect("missing source address");
+        let astr = matches.opt_str("a").expect("missing ammount");
+        let a = astr.parse().expect("ammount is not a number");
+        transfer(to, from, a);
+        return;
+    } else if matches.opt_present("b") {
+        let to = matches.opt_str("t").expect("missing destination address");
+        balance(to);
+        return;
     }
 }
